Move server data load job to services. Add sob starting to service initialiser
