+++ kiosk/src/js/services/events/internal-events-listeners/hotkeys_events_handler.js
@@ -1,6 +1,6 @@
 const electron = require('electron');
 const app = electron.app;
-const ServerDataLoadJob = require('../../../helpers/server_data_load_job');
+const ServerDataLoadJob = require('../../data/server_data_load_job');
 const {scheduledTaskManager} = require('../../../scheduled_task_manager');
 const WindowsHelper = require('../../windows/windows_helper');
